+++ CHANGELOG.md
@@ -20,8 +20,12 @@ Links to such platforms may or may not be valid.
 Releases are done everytime an important ticket is completed.
 ## [Unreleased]
 ### Major Changes
-* **2021-11-06**: Rename `net.splitcells.dem.data.order.Comparator.comparator_`
-  to `comparatorLegacy` in order to have a clear name.
+* **2021-11-06**:
+  1. **\#10** Remove `net.splitcells.dem.data.order.Comparator.comparator_`.
+     to `comparatorLegacy` in order to have a clear name and remove its usage,
+     where it is not needed.
+  2. **\#10** Rename `net.splitcells.dem.data.order.Comparators.comparators`
+     to `comparator` in order to unify naming.
 * **2021-11-05**: `system.update` for Flatpak does exit with an error,
   if Flatpak's version itself is deprecated.
   This is done because otherwise the whole update process would be aborted.